+++ docs/versionhistory.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 **UNRELEASED**
 
 - Added introspection of running tasks via ``anyio.get_running_tasks()``
+- Added the ``getsockopt()`` and ``setsockopt()`` methods to the ``SocketStream`` API
 - Fixed mishandling of large buffers by ``BaseSocket.sendall()``
 
 **1.0.0b1**